New SequentialActionSequence is used from Arbitraries
